Notifications / bidirectional transports
Using a router on the client side, this PR adds initial support for
bi-directional communication over transports that support it (sockets /
websockets).

Along the way, we also gain support for notifications, replacing the
current `onProcessMessage` hook which currently is used as a hack for
reading notifications in the client but doesn't offer full
bi-directionality. The processing hook will eventually get removed since
it complicates the implementation.

* add `notify` which sends a JSON-RPC message without the expectation of
a response
* deprecate `onProcessMessage` - replaced by `router`
* normalise logging topic style
* normalise max message size across clients/server