Add custom framing protocols for socket client (#253)
When using the socket client, framing defines how to separate one
message from another.

Currently, "\r\n" has been the only option, but it has some problems
with newlines in the payload (not a problem when using tightly encoded
json, but can happen with custom payload encodings) and with the size of
the message not being known up-front.

This PR adds two more options, both based on the .NET JsonRpc
implementation (for interoperability with vscode, among other things):

* "http-style headers": `Content-Length: length` prefix
* binary length prefix: Big-endian 32-bit length prefix