Cross-thread RPC via async channel
RpcChannel allows using a pair of [async
channels](https://github.com/status-im/nim-async-channels) to
communicate between threads as if they were connected over a socket or
other transport.

This setup avoids having to copy the data onto the OS buffer and instead
uses a ThreadSignalPtr to wake the target thread.

While it is not the most performant way to make cross-thread calls, it
is quite convenient for services that already expose a JSON-RPC server -
the choice of an internal channel then becomes a simple configuration
option where the only thing that changes is the setup.